+++ apps/foh/src/state/championships-state.ts
@@ -1,6 +1,10 @@
 import { atom } from 'jotai';
-import { Championship, collection } from 'lib';
+import { Championship, collection, filter, orderByDesc } from 'lib';
 
 export const championshipsState = atom(() =>
-  collection<Championship>('championships').get()
+  collection<Championship>(
+    'championships',
+    filter('published', '==', true),
+    orderByDesc('nr')
+  ).get()
 );