+++ packages/lib/src/index.ts
@@ -3,3 +3,4 @@ export * from './model/championship';
 
 // Repository
 export * from './repository/collection';
+export * from './repository/constraint';