+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lib
 
+## 0.0.5
+
+### Patch Changes
+
+- Add collection constraints.
+
 ## 0.0.4
 
 ### Patch Changes